Updated the dismissal to work on phones. More info...
Updated to allow you to see what user left on mobile devices. On Android I noticed this problem, the update just adds an extra piece that shows the display name along with the @invalid_user label. (Not sure about iOS, I don't use it, but it's there now just in case!)

Also updated to start the process of completely separating the text typos so it's different from the welcome cog!